% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predictions_to_db.R
\name{add_model_id}
\alias{add_model_id}
\title{Add model ID}
\usage{
add_model_id(model_descr)
}
\arguments{
\item{model_descr}{data frame}
}
\value{
same data frame where the new ID has been written into the model_id column
}
\description{
If a model with the same date_time, model_name, and description already exists in
the database, its ID is used. If not, the smallest ID that's not in the database is added.
}
