% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_modeling.R
\name{best_hyperparams_to_json}
\alias{best_hyperparams_to_json}
\title{Storing optimal set of HP}
\usage{
best_hyperparams_to_json(cv_results, target_variable)
}
\arguments{
\item{cv_results}{List containing sublists of all HP combinations}

\item{target_variable}{Either "q_kfz" or "v_kfz"}
}
\description{
Saves best set of HP as .json. It can be easily loaded from \code{inst/params}, where it is stored immediately after running the model.
}
