% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_modeling.R
\name{compute_avg_n_det}
\alias{compute_avg_n_det}
\title{Get average number of unique detectors per hour}
\usage{
compute_avg_n_det(dat)
}
\arguments{
\item{dat}{data.frame with columns date_time, x, and y}
}
\description{
For temporal CV we need the number of detectors per hour to multiply it with
the window width etc.
This number varies from hour to hour, however this won't affect the results very much so we
just use the average number of detectors per hour.
}
