% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_send_data_to_prod.R
\name{copy_data_to_prod}
\alias{copy_data_to_prod}
\title{Send data from dev schema to prod schema}
\usage{
copy_data_to_prod(x)
}
\arguments{
\item{x}{(list) list obtained from single_source, containing the fields table (the table to transfer),
source_schema (the schema to get the data), target_schema  (the schema to insert the data)
optimizes dubplicates}
}
\description{
Send data from dev schema to prod schema
}
