% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_env.R
\docType{data}
\name{env}
\alias{env}
\title{env module controlling dev or prod envrironment for db
if ENV is not set, it will be set to DEV
if ENV is set to PROD, the "prod_" prefix will be returned to be used at db}
\format{
An object of class \code{module} (inherits from \code{list}) of length 1.
}
\usage{
env
}
\description{
env module controlling dev or prod envrironment for db
if ENV is not set, it will be set to DEV
if ENV is set to PROD, the "prod_" prefix will be returned to be used at db
}
\keyword{datasets}
