% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_modeling.R
\name{folds}
\alias{folds}
\title{Make Folds for CV}
\usage{
folds(dat_train, n_splits = 4, n_weeks = 12)
}
\arguments{
\item{dat_train}{Training Date you want to split in folds}

\item{n_splits}{Number of splits}

\item{n_weeks}{The time range shifting with every new split}
}
\description{
Creates CV-Folds for usage in \code{xgb.cv}. The default can be changed to match your specific needs.
}
