% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{format_date}
\alias{format_date}
\title{Format date as character
e.g., 2022-07-29 01:00:00}
\usage{
format_date(date)
}
\arguments{
\item{date}{Date or PosixCT}
}
\description{
Format date as character
e.g., 2022-07-29 01:00:00
}
