% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_get_data.R
\name{get_data}
\alias{get_data}
\title{Get data for modelling}
\usage{
get_data(target_variable, dev, sample_fract = 0.002)
}
\arguments{
\item{target_variable}{(character) target value for model.
"q_kfz" for counted traffic of all vehicles volume/quantities,
"v_kfz" for average traffic speed of all vehicles.}

\item{dev}{(bool) If TRUE we work with less cases and less hyper parameter
combinations. Usually filled via Sys.getenv("DEV")}

\item{sample_fract}{(numeric) share of data to be used via sampling for model}
}
\value{
(data.frame) df that contains target variable and all predictors
}
\description{
Get data for modelling
}
