% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predictions_to_db.R
\name{get_latest_model_id}
\alias{get_latest_model_id}
\title{Retrieve latest model ID from database}
\usage{
get_latest_model_id(target_variable)
}
\arguments{
\item{target_variable}{(string) name of the target variable: q_kfz or v_kfz}
}
\value{
latest model ID with available model_object
}
\description{
Retrieve latest model ID from database
}
