% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_logging.R
\name{logging}
\alias{logging}
\title{Logging function}
\usage{
logging(msg, ...)
}
\arguments{
\item{msg}{The message to log}

\item{...}{Additional arguments to pass to the logging function}
}
\description{
Logging function
}
