% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_modeling.R
\name{make_DMatrix}
\alias{make_DMatrix}
\title{Write data as DMatrix}
\usage{
make_DMatrix(dat, features, target)
}
\arguments{
\item{dat}{Data to be rewritten}

\item{features}{Features to be included}

\item{target}{Targeted variable}
}
\description{
Uses \code{xgb.DMatrix} from the \code{xgboost} Package to transform data to the correct format
}
