% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_modeling.R
\name{make_predictions.default}
\alias{make_predictions.default}
\title{Make predictions using a given model object}
\usage{
\method{make_predictions}{default}(dat, model_object, target_variable)
}
\arguments{
\item{dat}{data frame with the features}

\item{model_object}{a single xgb model_object, e.g. xgb_fit$final model or
an model object, retrieved from the database with retrieve_model_from_db()}

\item{target_variable}{(character) "q_kfz" or v_kfz"}
}
\value{
vector of predictions
}
\description{
Make predictions using a given model object, e.g., from the database
}
