% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_modeling.R
\name{make_predictions.xgb.DMatrix}
\alias{make_predictions.xgb.DMatrix}
\title{Make predictions using a given model object with data already converted to xgb.Dmatrix}
\usage{
\method{make_predictions}{xgb.DMatrix}(dat, model_object, target_variable)
}
\arguments{
\item{dat}{xgb.DMatrix}

\item{model_object}{a single xgb model_object, e.g. xgb_fit$final model or
an model object, retrieved from the database with retrieve_model_from_db()}

\item{target_variable}{other paramters passed to make_predictions}
}
\value{
vector of predictions
}
\description{
Make predictions using a given model object, e.g., from the database
}
