% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predictions_to_db.R
\name{model_description_for_db}
\alias{model_description_for_db}
\title{Prepare model description for DB}
\usage{
model_description_for_db(xgb_fit, target_variable, dat_train, dat, model_name)
}
\arguments{
\item{xgb_fit}{XGBoost model object (to extract hyper parameters)}

\item{target_variable}{(character) "q_kfz" for counted traffic of all
vehicles volume/quantities, "v_kfz" for average traffic speed of all vehicles}

\item{dat_train}{(data.frame) training data (to extract beginning and end of training period)}

\item{dat}{(data.frame) full data (to extract date when the model was developed)}

\item{model_name}{(string) A name for the model, e.g., "without_weather"}
}
\value{
data.frame with columns model_id, date_time, model_name, depvar, and description
}
\description{
Prepare model description for DB
}
