% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predictions_to_db.R
\name{model_description_json}
\alias{model_description_json}
\title{Prepare json string with model description}
\usage{
model_description_json(xgb_fit, target_variable, train_date_time_column)
}
\arguments{
\item{xgb_fit}{XGBoost model object (to extract hyper parameters)}

\item{target_variable}{(character) "q_kfz" for counted traffic of all
vehicles volume/quantities, "v_kfz" for average traffic speed of all vehicles}

\item{train_date_time_column}{datetime vector of training data (to extract beginning and end of
training period)}
}
\value{
json string
}
\description{
Prepare json string with model description
}
