% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_modeling.R
\name{model_formula}
\alias{model_formula}
\title{Latest model formula}
\usage{
model_formula(target_variable)
}
\arguments{
\item{target_variable}{(character) "q_kfz" for counted traffic of all
vehicles volume/quantities, "v_kfz" for average traffic speed of all vehicles}
}
\value{
model formula
}
\description{
Latest model formula
}
