% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_modeling.R
\name{optimal_hyper_parameters}
\alias{optimal_hyper_parameters}
\title{Optimal hyper parameters}
\usage{
optimal_hyper_parameters(target_variable)
}
\arguments{
\item{target_variable}{(character) "q_kfz" for counted traffic of all
vehicles volume/quantities, "v_kfz" for average traffic speed of all vehicles}
}
\value{
data frame with optimal hyper parameters
}
\description{
As identified using temporal CV
}
