% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_modeling.R
\name{optimal_lambda}
\alias{optimal_lambda}
\title{Optimal lambdas}
\usage{
optimal_lambda(target_variable)
}
\arguments{
\item{target_variable}{(character) "q_kfz" for counted traffic of all
vehicles volume/quantities, "v_kfz" for average traffic speed of all vehicles}
}
\value{
integer with best lambda values, identified for the above hyper parameters
}
\description{
As identified using temporal CV
}
