% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_model_plots.R
\name{plot_pred_vs_obs}
\alias{plot_pred_vs_obs}
\title{Plot predicted against observed values}
\usage{
plot_pred_vs_obs(target_variable, dat_test)
}
\arguments{
\item{target_variable}{(character) "q_kfz" for counted traffic of all
vehicles volume/quantities, "v_kfz" for average traffic speed of all vehicles}

\item{dat_test}{df with columns pred and target_variable}
}
\description{
Plot predicted against observed values
}
