% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predictions_to_db.R
\name{preds_to_db}
\alias{preds_to_db}
\title{Retrieve data, make predictions, write them to DB}
\usage{
preds_to_db(
  query,
  xgb_fit,
  target_variable,
  model_id,
  table_name,
  database = Sys.getenv("DB_SCHEMA_TARGET")
)
}
\arguments{
\item{query}{(string) name of the query to retrieve the feature data}

\item{xgb_fit}{A model object from XGBoost}

\item{target_variable}{(string) name of the target variable}

\item{model_id}{(int) Model ID of the model we use}

\item{table_name}{(string) Name of the table where the predictions are sent}

\item{database}{(string) name of database schema the data is sent to}
}
\value{
Nothing
}
\description{
Retrieve data, make predictions, write them to DB
}
