% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predictions_to_db.R
\name{prep_preds_for_db}
\alias{prep_preds_for_db}
\title{Prepare data frame with predictions for DB}
\usage{
prep_preds_for_db(df, model_id)
}
\arguments{
\item{df}{data frame with predictions (columns date_time, x, y, value)}

\item{model_id}{model ID; may be NULL (then there won't be a column model_id)}
}
\value{
data.frame with columns model_id (if provided), date_time, x, y, value
in exactly this order
}
\description{
Add model ID, date_time_forecast and select columns
}
