% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_modeling.R
\name{random_hpo_grid}
\alias{random_hpo_grid}
\title{Grid for HPO random search}
\usage{
random_hpo_grid(tune_length)
}
\arguments{
\item{tune_length}{(int) number of rows for the grid, i.e., number of unique combinations
of hyper parameters}
}
\value{
data frame with random combination of hyper parameters
}
\description{
Create a grid of hyper parameters with random values.
We use this instead of the built-in random search of caret because we
want to specify the distributions to draw from.
n_rounds is a fixed, large number because the function is meant to be
used in combination with early stopping.
}
