% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predictions_to_db.R
\name{retrieve_model_from_db}
\alias{retrieve_model_from_db}
\title{Retrieve model object from database}
\usage{
retrieve_model_from_db(model_id)
}
\arguments{
\item{model_id}{(int) Model ID of the model we use}
}
\value{
model object
}
\description{
Retrieve model object from database
}
