% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predictions_to_db.R
\name{send_model_to_db}
\alias{send_model_to_db}
\title{Send model object to database}
\usage{
send_model_to_db(
  model_object,
  model_id,
  model_filename = "model.xgb",
  delete_local_copy = TRUE
)
}
\arguments{
\item{model_object}{A final model object from XGBoost, e.g. xgb_fit$finalModel}

\item{model_id}{(int) Model ID of the model we use}

\item{model_filename}{(string) model filename, model.xgb by default}

\item{delete_local_copy}{(logical) delete local copy of the model file after
sending it to the DB, TRUE by default}
}
\description{
Send model object to database
}
