% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_sources.R
\name{data_sources}
\alias{data_sources}
\alias{single_source}
\title{Data sources to query from clickhouse}
\usage{
data_sources()

single_source(
  table,
  source_schema = "fairq_features",
  target_schema = "fairq_prod_features"
)
}
\arguments{
\item{table}{(character) table on the database which data should be transferred.
has to exist in source and target schema.}

\item{source_schema}{(character) schema to get the data from}

\item{target_schema}{(character) schema to send the data to}
}
\description{
List data sources. This list defines which data source to transfer from dev to prod.
}
