% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_send_data_to_prod.R
\name{transfer_tables_to_prod}
\alias{transfer_tables_to_prod}
\title{transfer all tables to the prod schema}
\usage{
transfer_tables_to_prod(tables = data_sources())
}
\arguments{
\item{tables}{(list) see result of \code{data_sources()}}
}
\description{
transfer all tables to the prod schema
}
