% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_predictions_to_db.R
\name{preds_chunkwise_to_db}
\alias{preds_chunkwise_to_db}
\title{Retrieve data, make predictions on whole grid, write them to DB - chunkwise}
\usage{
preds_chunkwise_to_db(
  chunk_size,
  query_features,
  xgb_fit,
  model_id,
  table_name
)
}
\arguments{
\item{chunk_size}{(int) Size of the chunks, i.e., how many x coordinates are processed at once}

\item{query_features}{(string) name of the query to retrieve the features,
must be a query that takes the parameter x to filter the data on a specific x
coordinate. If the query is not parametrized by x, but retrieves all data at once,
use the function preds_to_db() instead of this one.}

\item{xgb_fit}{A model object from XGBoost}

\item{model_id}{(int) Model ID of the model we use}

\item{table_name}{(string) Name of the table where the predictions are sent}
}
\value{
Nothing
}
\description{
Processes the data chunkwise (by x coordinates) to save working memory.
Uses mode insert and optimizes table only once at the end to save time
}
